Use more appropriate data type

+++ capi/src/ffi.rs
@@ -1,4 +1,4 @@
 #[no_mangle]
-pub extern "C" fn on_key(code: u32) {
+pub extern "C" fn on_key(code: u8) {
     log::debug!("{code}");
 }
